# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API для управления подписками",
        "title": "Subscription Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/createSub": {
            "post": {
                "description": "Создает новую подписку. Если не указывать user_id он сгенерируется автоматически.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Создать новую подписку",
                "parameters": [
                    {
                        "description": "Данные подписки",
                        "name": "subscription",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestExamples.CreateSubscriptionReqExample"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Подписка создана",
                        "schema": {
                            "$ref": "#/definitions/responseExamples.SubCreatedExample"
                        }
                    },
                    "400": {
                        "description": "Пустая дата начала и дата окончания подписки или дата начала больше даты окончания",
                        "schema": {
                            "$ref": "#/definitions/responseExamples.EmptyDateOrInvalidExample"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/responseExamples.SomeServerErrorsDuriingCreatingSubExample"
                        }
                    }
                }
            }
        },
        "/deleteSub": {
            "delete": {
                "description": "Удаляет подписку по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Удалить подписку",
                "parameters": [
                    {
                        "description": "Данные для удаления",
                        "name": "subscription",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestExamples.DeleteSubscriptionReqExample"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Подписка удалена",
                        "schema": {
                            "$ref": "#/definitions/responseExamples.SubDeletedExample"
                        }
                    },
                    "404": {
                        "description": "Подписка не найдена",
                        "schema": {
                            "$ref": "#/definitions/responseExamples.SubNotFoundExample"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/responseExamples.SomeServerErrorsDuriingDeletingSubExample"
                        }
                    }
                }
            }
        },
        "/getAllSubs": {
            "get": {
                "description": "Получает список всех подписок",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Получить все подписки",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responseExamples.GetSudscriptionExample"
                            }
                        }
                    },
                    "404": {
                        "description": "Подписка не найдена",
                        "schema": {
                            "$ref": "#/definitions/responseExamples.SubNotFoundExample"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/responseExamples.SomeErrorGetingAllSubsExample"
                        }
                    }
                }
            }
        },
        "/getFullSubPriceByPeriod": {
            "post": {
                "description": "Подсчитывает общую стоимость и количество подписок пользователя за указанный период",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Получить суммарную стоимость подписок за период",
                "parameters": [
                    {
                        "description": "Параметры периода и пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestExamples.GetFullPriceByPeriodReqExample"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseExamples.GetFullPriceByPeriodExample"
                        }
                    },
                    "400": {
                        "description": "Пустая дата начала и дата окончания подписки или дата начала больше даты окончания",
                        "schema": {
                            "$ref": "#/definitions/responseExamples.EmptyDateOrInvalidExample"
                        }
                    },
                    "404": {
                        "description": "Подписка не найдена",
                        "schema": {
                            "$ref": "#/definitions/responseExamples.SubNotFoundExample"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/responseExamples.SomeServerErrorsDuringGettingSubExample"
                        }
                    }
                }
            }
        },
        "/getSub/{subscriptionID}": {
            "get": {
                "description": "Получает информацию о конкретной подписке",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Получить подписку по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID подписки",
                        "name": "subscriptionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseExamples.GetSudscriptionExample"
                        }
                    },
                    "400": {
                        "description": "Не указан ID подписки или ID подписки не является числом",
                        "schema": {
                            "$ref": "#/definitions/responseExamples.EmptySubOrInvalidIDExample"
                        }
                    },
                    "404": {
                        "description": "Подписка не найдена",
                        "schema": {
                            "$ref": "#/definitions/responseExamples.SubNotFoundExample"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при получении подписки",
                        "schema": {
                            "$ref": "#/definitions/responseExamples.SomeServerErrorsDuringGettingSubExample"
                        }
                    }
                }
            }
        },
        "/updateSub": {
            "put": {
                "description": "Обновляет существующую подписку",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Обновить подписку",
                "parameters": [
                    {
                        "description": "Обновленные данные подписки",
                        "name": "subscription",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestExamples.UpdateSubscriptionReqExample"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Подписка обновлена",
                        "schema": {
                            "$ref": "#/definitions/responseExamples.SubUpdatedExample"
                        }
                    },
                    "400": {
                        "description": "Пустая дата начала и дата окончания подписки или дата начала больше даты окончания",
                        "schema": {
                            "$ref": "#/definitions/responseExamples.EmptyDateOrInvalidExample"
                        }
                    },
                    "404": {
                        "description": "Подписка не найдена",
                        "schema": {
                            "$ref": "#/definitions/responseExamples.SubNotFoundExample"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/responseExamples.SomeServerErrorsDuriingUpdatingSubExample"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "requestExamples.CreateSubscriptionReqExample": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "2023-12-31T00:00:00Z"
                },
                "price": {
                    "type": "integer",
                    "example": 100
                },
                "service_name": {
                    "type": "string",
                    "example": "Netflix"
                },
                "start_date": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426655440000"
                }
            }
        },
        "requestExamples.DeleteSubscriptionReqExample": {
            "type": "object",
            "properties": {
                "subscription_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "requestExamples.GetFullPriceByPeriodReqExample": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "2023-12-31T00:00:00Z"
                },
                "start_date": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426655440000"
                }
            }
        },
        "requestExamples.UpdateSubscriptionReqExample": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "2023-12-31T00:00:00Z"
                },
                "price": {
                    "type": "integer",
                    "example": 100
                },
                "service_name": {
                    "type": "string",
                    "example": "Netflix"
                },
                "start_date": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "subscription_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "responseExamples.EmptyDateOrInvalidExample": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "start_date and end_date are required or start_date must be before end_date"
                }
            }
        },
        "responseExamples.EmptySubOrInvalidIDExample": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Subscription ID is required or Invalid subscription ID"
                }
            }
        },
        "responseExamples.GetFullPriceByPeriodExample": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 2
                },
                "total_cost": {
                    "type": "integer",
                    "example": 1000
                },
                "user_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426655440000"
                }
            }
        },
        "responseExamples.GetSudscriptionExample": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "2023-12-31T00:00:00Z"
                },
                "price": {
                    "type": "integer",
                    "example": 100
                },
                "service_name": {
                    "type": "string",
                    "example": "Netflix"
                },
                "start_date": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426655440000"
                }
            }
        },
        "responseExamples.SomeErrorGetingAllSubsExample": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "some server errors during getting all subscriptions"
                }
            }
        },
        "responseExamples.SomeServerErrorsDuriingCreatingSubExample": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "some server errors during creating subscription"
                }
            }
        },
        "responseExamples.SomeServerErrorsDuriingDeletingSubExample": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "some server errors during deleting subscription"
                }
            }
        },
        "responseExamples.SomeServerErrorsDuriingUpdatingSubExample": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "some server errors during updating subscription"
                }
            }
        },
        "responseExamples.SomeServerErrorsDuringGettingSubExample": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "some server errors during getting subscription"
                }
            }
        },
        "responseExamples.SubCreatedExample": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Subscription created"
                }
            }
        },
        "responseExamples.SubDeletedExample": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Subscription deleted"
                }
            }
        },
        "responseExamples.SubNotFoundExample": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Subscription not found"
                }
            }
        },
        "responseExamples.SubUpdatedExample": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Subscription updated"
                }
            }
        }
    }
}